Move SignalGuard into cfg-selected signal imp module
The public SignalGuard carried its own unix vs not-unix #[cfg] split inline
in platform_abi's lib.rs, colliding namespace-wise with process_unix's private
guard of the same name. Relocate it into signal_{unix,other}.rs selected by
cfg_attr path, matching the existing process module; lib.rs keeps only the
documented re-export.
