+++ src/platform_abi/src/signal_other.rs
@@ -1,0 +1,9 @@
+#[derive(Default)]
+pub struct SignalGuard;
+
+impl SignalGuard {
+    #[must_use]
+    pub fn new() -> Self {
+        Self
+    }
+}
